requirements.txt
Flask
flask-cors
psutil
Pillow
